Homepage: Refactor router into App
