+++ apps/homepage/src/index.js
@@ -1,23 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './pages/App/App';
-import About from './pages/About/About';
+import App from './App/App';
 import reportWebVitals from './reportWebVitals';
 import store from './store/store';
 import { Provider } from 'react-redux';
-import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <Provider store={store}>
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<App />} />
-        <Route path="/:lang" element={<App />} />
-        <Route path="/about" element={<About />} />
-        <Route path="/:lang/about" element={<About />} />
-      </Routes>
-    </BrowserRouter>
+    <App />
   </Provider>
 );
