--- apps/homepage/src/pages/App/App.test.js
@@ -1,8 +1,0 @@
-// import { render, screen } from '@testing-library/react';
-// import App from './App';
-
-test('Needs tests', () => {
-  // render(<App />);
-  // const linkElement = screen.getByText(/learn react/i);
-  // expect(linkElement).toBeInTheDocument();
-});
